requirements.txt
streamlit==1.3.1
requests==2.26.0
matplotlib==3.4.3
pandas==1.3.3
pyrebase4==4.5.0
plotly